Files to display posts in a topic
Attempt to create a .html and .js file to view the posts in a topic. I
am not sure if this is the route we will end up going. Currently, I
think the sql statement needs to be changed and the ajax auto update
does not work right. Also, I am unsure how to get the thread id from the
gettopic.php file to the topic.php file.